refactor(resources): Resources no longer pulled from GH
To further make Iris local-first, Resources will no longer be
pulled from the GitHub website. Instead, they're expected to be
loaded via a Data Pack.
