+++ src-tauri/src/resources/mod.rs
@@ -1,4 +1,3 @@
 pub mod commands;
-mod helper;
 pub mod schema;
 pub mod structs;
