+++ src/app/resources/components/ResourcesList.module.css
@@ -3,8 +3,7 @@
   flex-direction: row;
 
   background-color: #cbd5e0;
-  width: 90%;
-  max-width: 650px;
+  width: 650px;
 
   border-radius: 10px;
   padding: 5px 20px;
@@ -50,9 +49,9 @@
 
 @keyframes fadeIn {
   from {
-      opacity: 0%;
+    opacity: 0%;
   }
   to {
-      opacity: 100%;
+    opacity: 100%;
   }
 }
